feat: Users CRUD
- renamed userController to usersController
- added getById, index and delete
