--- app/Http/Controllers/UserController.php
@@ -1,33 +1,0 @@
-<?php
-
-namespace App\Http\Controllers;
-
-
-use Illuminate\Http\Request;
-use App\Models\User;
-use Illuminate\Support\Facades\Hash;
-
-class UserController extends Controller
-{
-    public function store(Request $request)
-    {
-        $request->validate([
-            'name' => 'required|string|max:255',
-            'email' => 'required|string|email|max:255|unique:users',
-            'password' => 'required|string|min:8|confirmed',
-            'phone' => 'required|string|max:15',
-            'role_id' => 'required|integer'
-        ]);
-
-        $user = User::create([
-            'name' => $request->name,
-            'email' => $request->email,
-            'password' => Hash::make($request->password),
-            'phone' => $request->phone,
-            'role_id' => $request->role_id
-        ]);
-
-
-        return response()->json(['message' => 'User registered successfully'], 201);
-    }
-}
